Allow overriding the CMake binary used for testing
This is useful when cross-compiling: In my case I run the tests on a
QEMU VM where CMake is not installed to the same path as the host
CMake used to build libpng. This commit adds a new TEST_CMAKE_COMMAND
variable that can be set to override the path of the CMake binary used
to drive the test.